refactor keystatic UI to use links in breadcrumbs
